+++ kotlin-android/src/main/java/com/segment/analytics/kotlin/android/plugins/AndroidContextPlugin.kt
@@ -1,4 +1,4 @@
-package com.segment.analytics.platform.plugins.android
+package com.segment.analytics.kotlin.android.plugins
 
 import android.Manifest.permission
 import android.annotation.SuppressLint
