+++ kotlin-android/src/main/java/com/segment/analytics/kotlin/android/utilities/AndroidKVS.kt
@@ -1,11 +1,7 @@
-package com.segment.analytics.utilities
+package com.segment.analytics.kotlin.android.utilities
 
 import android.content.SharedPreferences
 import com.segment.analytics.kotlin.core.utilities.KVS
-import com.segment.analytics.kotlin.core.utilities.createDirectory
-import java.io.File
-import java.io.FileOutputStream
-import java.time.Instant
 
 class AndroidKVS(val sharedPreferences: SharedPreferences): KVS {
     override fun getInt(key: String, defaultVal: Int): Int =
