+++ samples/kotlin-android-app-destinations/src/main/java/com/segment/analytics/destinations/plugins/MixpanelDestination.kt
@@ -7,10 +7,10 @@ import com.mixpanel.android.mpmetrics.MixpanelAPI
 import com.segment.analytics.kotlin.core.*
 import com.segment.analytics.kotlin.core.platform.DestinationPlugin
 import com.segment.analytics.kotlin.core.platform.plugins.LogType
-import com.segment.analytics.platform.plugins.android.AndroidLifecycle
+import com.segment.analytics.kotlin.android.plugins.AndroidLifecycle
+import com.segment.analytics.kotlin.android.utilities.toJSONObject
 import com.segment.analytics.kotlin.core.platform.plugins.log
 import com.segment.analytics.kotlin.core.utilities.*
-import com.segment.analytics.utilities.*
 import kotlinx.serialization.json.*
 import java.util.*
