+++ samples/kotlin-android-app/src/main/java/com/segment/analytics/next/plugins/PushNotificationTracking.kt
@@ -2,10 +2,9 @@ package com.segment.analytics.next.plugins
 
 import android.app.Activity
 import android.os.Bundle
-import android.util.Log
 import com.segment.analytics.kotlin.core.Analytics
 import com.segment.analytics.kotlin.core.platform.Plugin
-import com.segment.analytics.platform.plugins.android.AndroidLifecycle
+import com.segment.analytics.kotlin.android.plugins.AndroidLifecycle
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.put
